finished moving code into their respective library modules
