+++ haskell/dev_run.sh
@@ -1,3 +1,3 @@
-# rm -rf *pem client/ plugin/ server/ p2p/ p2prc.[pP]* config.json dist-newstyle/ &&\
+rm -rf *pem client/ plugin/ server/ p2p/ p2prc.[pP]* config.json dist-newstyle/ &&\
   cabal clean &&\
   cabal run
